requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.4.1 # https://github.com/andymccurdy/redis-py
maxminddb==1.5.2
maxminddb-geolite2==2018.703
docutils==0.15.2
draftjs-sanitizer==1.0.0

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/niwinz/django-redis
# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework

django-templated-email==2.3.0 

django-countries==6.1.2
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
django-versatileimagefield==2.0
django-celery-beat==2.0.0

django-graphql-jwt==0.3.1
django-graphql-auth==0.3.10
django-graphiql==0.4.4
django-filter==2.2.0

django-countries==6.1.2
django-measurement==3.2.0
django-mptt==0.11.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.0.2



# Graphql
#--------------------------------------------------------------------------------
graphene-django==2.9.1
graphene-federation


# Others
#--------------------------------------------------------------------------------
celery
google-i18n-address==2.3.5
opentracing==2.3.0 
sentry-sdk
 
